Add test.check spec for structure-digit encode
I want to use this as a baseline for an extraction of encode's
assertions into a new pred fn. I'm hoping that test.check won't find any
examples of strings that pass those assertions but still fail to encode
for some reason. If that all plays out, I'm going to use this to
optimize some code in a project that consumes this library.